[db] Add new UserData model that FKs to UserVault (#98)
Here, we add a new table called UserData that foreign keys into the UserVault table. Each row represents a singular piece of data for a user, like an email or an ssn. Each row can be "validated" independently from others.

This is beneficial to support the case when a user may have multiple email addresses or phone numbers. Eventually, it can also graduate to serve as an audit log of when users change their data - for example, if they update their last name, we could insert a new UserData row with the new last name and mark the old UserData row as "deactivated" and include a tombstone timestamp.

This PR only migrates emails to use this new table and does so transparently - we can do a similar migration for phone numbers/names/etc in the future.

commit-id:ea2be7ff