[ci] Retry eventually consistent integration tests until success (#249)
Some operations require waiting for an SMS message to be delivered to a
twilio phone number/rate limit timeout to be met. This adds a little
util that can be used to retry any operation for some number of seconds.

Also speeds up integration tests a lot when running locally because you can reduce the rate limit between sending SMS messages to the same phone number

feature-branch: ef/retry