[idv] More advanced name difference calculation (#746)
feature-branch: elliott/fp-1292-smarter-calculation-of-distance-between

This tries to be more forgiving for common discrepancies between the caller ID name and the name we have on file. The two big cases this protects against:
- Extra names (like middle names) in the caller ID. We now only care that the name on file is a subset of the caller ID name, if provided
- Out-of-order names (like `Last First` instead of `First Last`)

To accomplish this, we take all of the possible length-2 permutations of the names inside of the caller ID name. For example, if `ELLIOTT VETLE FORDE` were provided as the caller name, we would try the permutations `[ELLIOTT VETLE], [ELLIOTT FORDE], [VETLE ELLIOTT], [VETLE FORDE], [FORDE ELLIOTT], [FORDE VETLE]` against the name we have on file for the user, [ELLIOTT FORDE]. We only return the minimum levenshtein difference among all of these permutations.