[idv] Allow building UserVaultWrapper from VerificationRequest (#753)
This is currently a premature optimization: in the future, if there's a VerificationRequest with no response, we'll need to reconstruct the user's data at the time the request was created and send it off to data vendors again. Or, if the requests are made in a background process, this is required to fetch all of the user's data.

It's a useful experiment in the short term, so I figure we will keep it, but it's not really adding any functionality right now.

feature-branch: ef/uvw-build