[api] Use type system to require checking permissions on WorkOs auth (#765)
[api] Use type system to require checking permissions on tenant auth

feature-branch: elliott/fp-1315-use-type-system-to-enforce-permissions

Very similar to the refactor we did for user auth, we now:
1. Parse the request into a SessionContext<ParsedWorkOs>, which has no utility functions to get information on the tenant
2. Require that the user call `.check_permissions(vec![...])` on this `SessionContext<ParsedWorkOs>` in order to get a `SessionContext<WorkOs>`. Only the latter struct implements the utility functions on the `TenantAuth` trait that allows the endpoint to do anything useful with the auth.

Notably, we have to create a `CheckTenantPermissions` trait for tenant auth, and the `check_permissions()` function has to return a `Box<dyn TenantAuth>`. This is because there are multiple types of tenant auths (secret key auth + workos session auth) that require permission checking. We use `Either<>` for tenant auth classes pretty frequently, but we never use it for user auth classes since there is only one