[api] Create new /onboarding/authorize endpoint (#830)
feature-branch: elliott/fp-1402-post-hostedonboardingauthorize

Repurposes the old `POST /hosted/onboarding/complete` endpoint into the new `POST /hosted/onboarding/authorize` endpoint. The purpose of this endpoint is now much more explicit: it doesn't initiate the KYC check, it solely marks the onboarding as authorized, giving the tenant permissions to this user's data. It also generates a validation token that the tenant can use to auth the user