[api] Don't allow deactivating the currently authed tenant user (#903)
feature-branch: elliott/fp-1492-dont-allow-deactivating-the-currently

## Describe your changes
Add-on to the new endpoint added in #900. This will help us avoid an
annoying case where the only user of a tenant deactivates their own user
and can't get back into the tenant

- Creates a version of the `check_permissions` and `can_decrypt`
functions that don't return the generified `Box<dyn TenantAuth>` trait
objects. These trait objects lose information about the underlying type
that implements the TenantAuth trait, so we aren't actually able to get
information on the authed `TenantUser`.
- Uses this new `check_permissions` function to get a version of the
tenant auth that can share information on which user is authed. Make
sure we're not deactivating our own user