[api] Allow re-inviting deactivated TenantUsers (#904)
feature-branch: elliott/fp-1491-what-happens-when-we-try-to-invite-a

## Describe your changes
In #900, we added the ability to deactivate a user on the admin
dashboard. Since DB DELETEs are scary and lose audit history, this is
implemented by setting a "deactivated" flag on the TenantUser and not
allowing you to log into a "deactivated" TenantUser.

Now, if you "delete" (aka deactivate) a user, you are able to re-invite
them back to the dashboard by creating a new active TenantUser row. This
keeps an entire audit-able history of when a given user had access to a
tenant