[api] Create tenant role deactivate endpoint (#914)
feature-branch: elliott/fp-1358-tenantrole-deactivate-endpoint

## Describe your changes
Supports deactivating unused TenantRoles. We don't want to do a DB
delete operation because they're scary and we lose some audit trail
history.

This one is a little more tricky... When we deactivate a tenant role, we
want to make sure that there are no users that are using the role at the
time it's deactivated. To protect against this:
- When creating/updating a TenantUser to use a certain role, we now lock
the role to prevent it from being deactivated
- When deactivating a TenantRole, we lock the role and then check that
nobody is using the role while holding the lock

Though it's impossible by codepaths that exist today, we also add a
check when logging into a TenantUser that its TenantRole is active.

We'll now actually filter out the deactivated TenantRoles from the list
endpoint.

## Caveats
Not my favorite implementation - if we decide to one day scope
TenantApiKey access using TenantRoles, it's very possible we forget to
update those codepaths to prevent deactivating a TenantRole that is in
use by an API key. The least complex and least error prone
implementation is just to show a message before deactivating that it
will revoke permissions for anyone using the role - so it may be
possible for a live TenantUser or TenantApiKey to be using an inactive
role. But then we can also allow re-activating an inactivated role. This
also doesn't protect against rogue prod scripts that disable
TenantRoles. Lmk if you prefer this implementation and I can chat with
Pedro about updating the designs

After typing out the caveats, it makes me more strongly prefer the
alternative implementation....