[api] Update User serialization to better match spec (#1058)
feature-branch: ef/user


## Describe your changes
The spec
[here](https://www.notion.so/onefootprint/User-detail-Audit-Trail-API-schema-draft-8af1742247a849e5b8476a547d538856)
has some changes since this was implemented - just updating the
serialization of the ScopedUser to match the new spec.

This requires now fetching the `latest_decision` for an onboarding.

## Caveats
In the future, we'll have to fetch the `TenantUser` that made the
decision if an optional `tenant_user_id` exists on the
OnboardingDecision