[db] Add new UserTimeline model (#1087)
feature-branch: elliott/fp-1645-create-new-db-table

## Describe your changes
We are deprecating the old AuditTrail model in favor of a "Timeline" of
events that have happened to the user. Now that we have the risk signals
section to help explain why footprint made a certain decision, the
timeline no longer needs to serve that purpose - this new timeline is
very similar to the audit trail, but its purpose is more to provide
useful information on events over time rather than to justify why
footprint made a certain decision.

We'll support a few different types of events:
- Data has been collected (created in `POST /identity_data`
- Biometric credential has been registered
- Document has been uploaded
- Decision has been made (either by footprint or by an admin user at the
tenant)

Another notable difference in this model is that the Timeline events
generally point to the DB object that created the timeline event - when
a decision is made, we'll reference the OnboardingDecision in the DB,
which allows us to dynamically assemble content about the resource.
Maybe we'll get rid of the timeline table eventually and just interleave
events that occurred across different tables