[db] Start creating UserTimeline events in the db (#1091)
feature-branch: elliott/fp-1646-start-inserting-rows-into-timeline-table

## Describe your changes
This starts creating UserTimeline events when a biometric credential is
registered AND when a decision is created.

## Caveats
Notably, we don't yet create UserTimeline events for document upload
(code doesn't exist yet) or adding user data. The latter is more
challenging because we add phone numbers, emails, and identity data in 3
different places. The timeline would be a little clunky...

I'm almost thinking that we'll want to wait until we have a more fleshed
out data model for tracking who adds what to display these in the
timeline. We could create the timeline events just for tracking now if
we want to, but I think they'll be too ugly to display