[workflows] Update build_and_deploy to be able to deploy a specific SHA
And start passing in SHA from the prod deploy job. Other deploy jobs will just deploy the commit that triggered the workflow (which is easy for push triggers). The prod deploy is trickier because it is triggered by a successful dev deploy. So, the prod deploy job has some logic to extract the correct commit hash and pass it into the build_and_deploy workflow

More info: https://linear.app/footprint/issue/FP-1698/does-prod-deploy-triggered-by-dev-success-deploy-same-version-as-dev