[api] Serialize ob config must collect attributes in decision timeline event (#1233)
## Describe your changes
For the user timeline on the dashboard, we need to include which fields were submitted to vendors. This is the set of fields that was requested by the ob configuration.

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/4ebe6ab0-883d-4fb6-a732-6083e719e9fd/image.png)

Resolves FP-1781

## Caveats

Notably, we right now won't show that the ID card was sent off for verification if it was a step up. It's a little more complex to do this - can punt for now. Will be easier when we can easily map a VerificationRequest to the set of fields that was verified by the request

List of vendors consulted for the decision will come in a follow-up Pr