[api] Refactor code to create fixture onboarding statuses (#1268)
## Describe your changes

There was a bug that was making two OnboardingDecisions for each test
user (one in `create_final_decision` and another in
`create_test_fixture_data`). Now that we start showing these in the
UserTimeline in admin dashboard, it became more urgent to clean up the
codepath that makes test data.

It was very complex before because we had interleaved test codepaths
with production codepaths. Even for test users, we would call the very
core pieces of decision logic in
`build_verification_requests_and_checkpoint` and `create_final_decision`
even though most of those pieces of logic don't apply to test users.

This pr breaks up the `decide` function into to very clear code paths.
We now:

- First check if the user vault is a test user vault. If so, we
immediately create all the fixture VerificationRequest,
VerificationResult, OnboardingDecision, and UserTimeline events and
update the Onboarding's status to the desired status
- Otherwise, if the user vault is for production, create the
verification requests in the DB and update the onboarding status to
Processing
- When responses come back from vendors, call
`risk::create_final_decision`


Resolves FP-1840


This also removes the creation of the old `AuditTrail`s (the old
`UserTimeline` model), which coincidentally lets us simplify some code
and not have to pass around the ob_config or tenant_id.

## Caveats
Now there's a bigger branch between sandbox users and production users,
but i think that's better