[api] Remove OnboardingStatus and onboarding.status column (#1290)
## Describe your changes

Resolves FP-1856

This removes the `Onboarding.status` column. This column was always a
denormalized view that tracked a lot of different sources of truth, and
before the statuses v2 migration, it was used both as source-of-truth
book keeping and as a denormalized display column, which was extremely
confusing.

Now, an onboarding's "status" is a dynamic function of any outstanding
ManualReview row for the onboarding and the latest OnboardingDecision
row for the onboarding. Notably, we un-denormalized the `status` on the
onboarding so it is now dynamically computed from its source of truth.

## Caveats

`DecisionStatus` should be a subset of `VisibleOnboardingStatus` (by
removing the StepUpRequired variant). This will happen in a follow-up
"decisions v2" migrations :
https://linear.app/footprint/issue/FP-1851/refine-representation-of-decisions

Now that the onboarding status is not denormalized, it becomes more
difficult to filter for onboardings in a specific state. It's usually
not too complex - there's just some more code in the handler for the
`GET /users` page shown in the admin dashboard.

If we want to keep `status` denormalized, we have to:
- lock the onboarding + update its status when a decision is made (not
too bad)
- lock the onboarding + update its status when a manual review is
created (not too bad)
- lock the onboarding + update its status when a manual review is
cleared. this is trickier because we have to then compute the status
that was previously being hidden by the manual review

An alternative could be to keep a denormalized `is_in_manual_review` and
a denormalized `latest_decision_status` on the onboarding row and
dynamically compute the status from these two fields. Happy to make this
change if you prefer

Another update: the more i think about it, maybe we should separate
`is_in_manual_review` and `latest_decision_status` (whether or not they
are denormalized onto the onboarding table). Otherwise, there's a weird
side effect where filtering for failed onboardings in the admin
dashboard also shows manual review onboardings

Another another update: decided to separate these in #1291

## Test Plan


Tested reverse migration:
```
footprint_db=# select onboarding.status as ob_status, onboarding_decision.status decision_status from onboarding left join onboarding_decision on onboarding_decision.onboarding_id = onboarding.id where onboarding_decision.deactivated_at is null order by 1;
    ob_status     | decision_status
------------------+------------------
 failed           | fail
 failed           | fail
 processing       |
 step_up_required | step_up_required
 step_up_required | step_up_required
 step_up_required | step_up_required
 verified         | pass
 verified         | pass
(8 rows)
```




https://user-images.githubusercontent.com/21066341/201430233-749f470d-a515-43c1-b322-a6afaa1bb3b9.mov