[api] Serialize source of annotation in GET /annotations (#1302)
## Describe your changes

In case we want to show who created the note that is pinned at the top of the admin dashboard.

This also starts to standardize the way we serialize an "actor" who performed an action. We'll probably solidify this more in the future to be either footprint, a tenant API key, or a tenant admin user