[api] Filter on is_live when looking up onboardings (#1304)
## Describe your changes

Forgot to filter on this when this was added earlier. Not super risky in
this context since a WorkOs context is allowed to do whatever they want
(they send an `x-is-live` header), but could be dangerous for a tenant
secret API key where we forget to restrict a sandbox key to touching
sandbox data.

Don't have a quick fix in diesel for this, but worth thinking about over
time