[api] Update POST /decisions endpoint to clear manual reviews (#1306)
## Describe your changes

Adapts the newly added `POST /onboarding/<id>/decision` endpoint to also
clear any outstanding manual review for an onboarding.

This also allows the same endpoint to be used to clear any manual review
while keeping the same decision on the backend.

Resolves FP-1857

## Caveats

Right now, we don't show anything in timeline when a manual review is
cleared, but we should.

Maybe we create a manual decision in this case? Even though the ob
status doesn't change? Or we just make a UserTimeline event with a new
type