[api] Update POST /decisions to take in fp_user_id (#1321)
## Describe your changes

We were thinking we would migrate endpoints on the admin dashboard to be onboarding-specific rather than fp-user-id specific, and the `POST /decisions` endpoint was the first one we wrote that was onboarding-specific.

We decided to instead just make fp_user_id <> onboarding a one-to-one relationship (in #1320), so now we can continue to follow the standard of having fp_user_id in the URL path.

Resolves FP-1627