Fix Select Options container height (#1348)
The Options container in the Select component will show 5.5 rows to make
it more evident that there are more options with scroll. Removing margin
bottom of the container as well.
<img width="1178" alt="CleanShot 2022-11-18 at 11 57 36@2x"
src="https://user-images.githubusercontent.com/12160059/202759799-dcee9e19-ce8d-4f89-8504-671937db0188.png">

This PR does not include the fix of the background color of the options
on hover. Check https://github.com/onefootprint/monorepo/pull/1333 for
that fix.

Co-authored-by: Rafael Michels Motta <rafael@onefootprint.com>