[FP-1169] Fix address infinite loop (#1494)
## Describe your changes

Changes:
- Fixes a bug that would provoke an infinite loop when selecting an
address using a mobile device. The issue was caused by
[Popper](https://popper.js.org/) the library which helps us to position
the dropdown correctly.
- Add tests (the component didn't have before)
- The behavior should be all the same.

This was my 2nd PR related to this ticket (the last one I close it): I
tried to remove the `downshift` dependency, but `react-select` doesn't
work well in this case, as it has a different structure that doesn't fit
on our use-case, and wouldn't allow using navigation with the keyboard,
so I decided to stick with the `downshift` dependency. We probably need
to evaluate better in a second moment about `react-select` x `downshift`

Demo:


https://user-images.githubusercontent.com/446731/205903660-a8eb7407-49b2-4a40-b4ae-831e241620fd.mov


## Issue: FP-1169

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->