Uses npx vercel when calling set:dot-env (#1602)
## Describe your changes

We have a command to pull the env variables to use when developing
locally. This command uses `vercel` cli, which must be installed. I
noticed some folks didn't have the cli installed - which is installed by
default in our `onboarding.sh` script.

This PR will make this process a bit simpler, given it will use
`npx`(https://www.freecodecamp.org/news/npm-vs-npx-whats-the-difference/)
to install a local version of `vercel cli`, so it makes everything more
bulletproof.

Furthermore, it simplifies the usage, as it will request the user to log
in (in case the user isn't logged). Previously, if you were not logged
in, would break and not continue.

Added a doc page as well:
https://www.notion.so/onefootprint/Environment-variables-d5537566621446278118df4d39fb1c53

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

1. Go to the frontend repo: `monorepo/frontend`
2. Run `yarn set:dot-env`