Fix master build
https://onefootprint.slack.com/archives/C03BC0SJ9GD/p1671149754638909

Main problem is that we run integration tests that created multiple scoped users for a single tenant, and then when we tried to run down.sql, we could create a unique constraint on (user, tenant).

There's no easy way to delete one of the conflicting scoped users since there are so many related records...

Then, in up.sql, we had some problems creating and index in the same transaction in which we backfill. So I tricked diesel to commit the transaction it starts, then open a new one to create the indexes.