Remove tenantPk from all unnecessary places in frontend (#1629)
## Describe your changes

Over time, we have been updating the backend to not require the tenantPk
except for in the identify flow - once an auth token is generated, it is
linked to a tenant PK :)

This finally removes the tenantPk from most endpoints on the frontend. 

Resolves FP-1830

## Test Plan

Ran through the bifrost flow, including handoff skipping liveness and
handoff properly registering liveness.