Remove JSON from handoff URL (#1633)
## Describe your changes

Just so it's less clunky for the customer. Now, we encode the opener
inside the URL querystring rather than in a JSON-serialized anchor

Resolves FP-2158

## Test plan
Went through bifrost flow in a few different ways to verify handoff
works. This `opener` field controls whether or not we show a countdown
when a secondary tab is opened for liveness on mobile, so I verified
this codepath works:

Show with/without mobile opener:

https://user-images.githubusercontent.com/21066341/209027176-7d8ae4be-a9ec-4c57-a897-61f1ec778635.mov

Show from clicking button:

https://user-images.githubusercontent.com/21066341/209027230-74fb9088-1fb0-44cf-b43e-108a367990b5.mov