Remove unwrap (#1697)
## Describe your changes
https://onefootprint.slack.com/archives/C044HVC8UM8/p1672773727935689

This is a codepath that //should// be safe to call `.unwrap()`, but is a good lesson to always catch it with application logic.

I think here we were somehow getting a duplicate scoped user from a DB query (not surprised, the logic to fetch them has complex joins that could yield multiple results. When we assemble UVWs for the scoped user, there's only one per unique scoped user id. This `.remove` from the hashmap makes it so later iterations can't find the same UVW