Create ability to select which tenant you'd like to log into (#1681)
## Describe your changes

If a user's email address is linked to multiple tenants, we should let
the user pick which tenant they want to log into. This may happen if I
invite a user by email to a tenant, but their email already has an
existing tenant. This kind of mimics the AWS log-in, which also gives
you a choice of which AWS account you want to log into if your user has
access to
multiple:![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/bd248be1-f02e-4250-abfd-20c9992fda55/image.png)

Uses the two new endpoints to (1) list tenants that a user has access to
and (2) log into the specified tenant to build a new UI for picking
which tenant you'd like to log into.

## Caveats
No users in prod are currently in this case, so nobody will see this
branch of the login flow. Designs are not final or pedro-approved in any
way 😛

But, this is a good warm-up for super-admin dashboard which will allow a
footprint employee to log in as any tenant. I'm thinking super-admin
dashboard may be best built as a new tab of the dashboard that is only
visible to footprint employees, and we can update the backend endpoints
to allow converting a footprint TenantUser auth into an auth that works
for any tenant

## Test Plan

[Screen Recording 2022-12-23 at
16.26.32.mov](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/5c061052-b78a-4b05-97a5-6d4da51f49cf/Screen%20Recording%202022-12-23%20at%2016.26.32.mov)

Resolves FP-2236