Build tenant permissions AST (#1711)
## Describe your changes

No functionality at all is changed in this PR - we simply migrate
existing auth utils to be more extensible for the future. Eventually, we
want to support an easily readable, more complex representation of
permissions, like "allow calling this HTTP handler if this auth token
has the Users scope, Admin scope, OR a read-only scope." The final PR in
this stack, #1726, shows the benefits of this AST representation of
permissions

This also attempts establishe some auth language - naming is hard, so
feel free to suggest new names lol
- `permission`s are set per HTTP endpoint handler. A provided auth class
must have sufficient scope tp perform the requested permission
- `scope`s are embedded inside of an auth token and allow certain
actions. For example, a TenantUser's auth session may have scopes that
allow access to the `GET /users` endpoint. (This is currently a
conflicting name with the TenantPermission enum which is used both as a
"permission" and a "scope")
- `auth extractor`s are the actix extractors that represent a principal
(either TenantUser, TenantApiKey, or Either<TenantUser, TenantApiKey>)
is authenticated

### Permissions vs Scopes

In this PR, permissions <> scopes are still really intertwined. The
vision is that one day they will have more distinct purposes

Permissions are required to execute an HTTP handler. Types of
permissions include:
- `Any` permission - always allow access
- `CanDecrypt(Vec<DataLifetimeKind>)` permission - allows access if the
granted scopes allow acces to decrypt a CollectedDataOption that
contains these DLKs
- `TenantPermission` - (introduced in this #1723) very basic permissions
that have a one-to-one mapping to TenantScopes. For example, a token
with `TenantScope::Users` satisfies a `TenantPermission::Users`
permission
- `Or<Left, Right>` - some combination of the above that is met if
either Left or Right is met
- (maybe one day) `And`


Scopes are granted to an auth token. Types of scopes include:
```
pub enum TenantScope {
    /// A special scope that gives permission to perform all actions
    Admin,
    /// (one day) A special scope that gives permission to endpoints that only require a ReadOnly permission
    ReadOnly

    OnboardingConfiguration,
    ApiKeys,
    OrgSettings,
    Users,
    /// Allows decrypting all custom attributes
    DecryptCustom,
    /// Similarly to how we store permissions on an OnboardingConfiguration, we denote the set of
    /// decryptable fields with CollectedDataOption rather than DataLifetimeKind
    Decrypt(Vec<CollectedDataOption>),
    ManualReview,
}
```


## Caveats

Lots of other fun improvements / clean up in future PRs on this stack

This kind of advances tenant IAM past user auth token permission
management. Will likely revisit user auth in the future

## Test Plan

Tests re-instated in #1724