Create separate TenantPermission enum from TenantScope (#1723)
## Describe your changes

Resolves [this
comment](https://github.com/onefootprint/monorepo/pull/1711/files#r1061190484)
from the first PR in the stack.

The representations of permissions required to execute an HTTP handler
vs the scopes granted to a TenantUser are different.

Permissions are required to execute an HTTP handler. Types of
permissions include:
- `Any` permission - always allow access
- `CanDecrypt(Vec<DataLifetimeKind>)` permission - allows access if the
granted scopes allow acces to decrypt a CollectedDataOption that
contains these DLKs
- `TenantPermission` - (introduced in this #1723) very basic permissions
that have a one-to-one mapping to TenantScopes. For example, a token
with `TenantScope::Users` satisfies a `TenantPermission::Users`
permission
- `Or<Left, Right>` - some combination of the above that is met if
either Left or Right is met
- (maybe one day) `And`


Scopes are granted to an auth token. Types of scopes include:
```
pub enum TenantScope {
    /// A special scope that gives permission to perform all actions
    Admin,
    /// (one day) A special scope that gives permission to endpoints that only require a ReadOnly permission
    ReadOnly

    OnboardingConfiguration,
    ApiKeys,
    OrgSettings,
    Users,
    /// Allows decrypting all custom attributes
    DecryptCustom,
    /// Similarly to how we store permissions on an OnboardingConfiguration, we denote the set of
    /// decryptable fields with CollectedDataOption rather than DataLifetimeKind
    Decrypt(Vec<CollectedDataOption>),
    ManualReview,
}
```

## Test Plan

Tests coming in future PR!