Skip production frontend builds when we don't have frontend changes (#1731)
## Describe your changes

When merging a PR to production, we had a condition that would rebuild
every app, even if we had only backend changes. Occasionally, these
builds would fail.

I'm disabling these builds for now, given they won't produce any real
value for us and this will be better to make our CI faster. In the
future, if we rely on any type of generation from the backend, we can
reevaluate and trigger a frontend build if necessary.

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

To test this, I did a change on the backend and I observed whether this
would produce a build or not. It didn't build, as expected