[FP-2378] Improve organizations page (#1786)
## Describe your changes

As discussed via slack recently, we had a small issue with the org
selector. However, I went a bit further.

Changes:
1) Previously we were using a table to display the options. This wasn't
ideal and was causing border-radius issues. Refactored to just use
buttons, so the problem is gone and now is fully accessible
2) The code wasn't catching server errors (for fetching orgs and
selecting one). Now we display both errors, in case they happen
3) Also added an error handling in case the token for some reason is not
provided
4) Added test cases, for everything

Did a small refactoring to make it simpler too, using `data`, `loading`
and `error` folders. The behavior should be the same.

Demo:



https://user-images.githubusercontent.com/446731/211573475-595d38d8-9078-4aa4-b532-75ac65207ae9.mov



## Issue: FP-2378

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->