UserConsent data model + API (#1747)
## Describe your changes

- Adds `UserConsent` data model for recording a user giving us consent
to collect and use biometric data.
- For now, contains a `consent_language_text` string field which we can
just populate with the exact language we showed the user when they hit
`Allow`. We could also make a separate data model that stores versions
of our consent language and instead have this have a reference to that.
Open to ideas here if others have a stronger sense of our requirements
here!
- Also links to an `InsightEvent` row that was created from the request
that created the `UserConsent`. This associates an ip_address and other
request level details with the act of consent which may be useful.
- Adds `POST /hosted/user/consent` which we would call during Bifrost
when the user clicks `Allow` to strongly capture the act of user consent
by writing a row in `UserConsent` for that
`user_vault_id`/`scoped_user_id`. Currently accepts
`consent_language_text` as a param.

cc @belcedogru 

## Issue: FP-2320

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->