Create uvw.scoped_user_id_or_else util (#1821)
## Describe your changes

Just a fun shorthand for this commonly used chunk of code.

## Caveats

We only have this because we don't have type-safety that certain methods
can only be called on UVWs built with a scoped_user_id. It would be more
hairy to turn this runtime check into a compile-time check, but maybe we
will in the future