FootprintReasonCode mapping for RiskSignal's (#1834)
## Describe your changes

v0 RiskSignal feature we aligned on IRL:
- RiskSignal's aren't _explanations_ of an onboarding decision. They are
instead just a (fairly) opaque abstraction over enumerated signals
(reason codes) that we are getting from multiple vendors.
- We are going for very high granularity. This means we aren't really
deduping or condensing reason codes within a single vendor. Our main
concern is (1) deduping across vendors and (2) deciding which vendor
reason codes we do/do not want to forward into RiskSignal's
- Main use of RiskSignals is (1) eye candy and (2) eventually to be used
in manual reviews
- We will continue with the current approach of writing PG RiskSignal
rows at onboarding decision time. We recognize this means we'll probably
have some hairy migrations that we'll eventually need to do when we make
changes to FootprintReasonCode's but we probably don't have enough
clarity on how we actually want to map vendor reason codes and what we
are achieving through to get this first on our pass here anyway.
- An alternative we should contemplate in the future is if we really
need to write this data at onboarding decision time or we could just
generate RiskSignal's (or at least do the reason code mapping logic) on
the fly when requests to `GET ../risk_signals/` are made

This PR:
- Consolidates both Socure and Idology reason code enums into
`../src/reason_code/`
- Defines `vendor_reason_code_enum!` macro which gives us a clean way to
specify for each vendor reason code its serialization, description, and
the FootPrintReasonCode it maps to (or that we are choosing not to map
it).
- Added some missing IdologyReasonCode's
- Populates `FootprintReasonCode` with most reason codes from Idology.
For a handful that I determined as not relevant, I removed.
- Mapping for all IdologyReasonCode's into FootprintReasonCode
- For Socure, I mapped all the obvious ones that are currently used in
the v6 decision logic. I also added a handful of variants to
FootprintReasonCode that Socure needed that Idology hadn't already
seeded.

I have not yet burned through this set of menial tasks but can do so if
the structure here looks good:
- Mapping all ~400 Socure reason codes to FootprintReasonCode's
- Setting `description`, `scopes`, `severity` for our 100+ initial set
of FootprintReasonCode's


## Issue: FP-2246

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->