Reuse the same tenant scope permission-checking code to check ob config permissions (#1881)
## Describe your changes

Permission checking for decryptions has always had two stages:
(1) verfiy that the authed tenant principal has a role with scopes that
give permissions to decrypt. These were checked with "guards"
(2) verify that the user has authorized an onboarding configuration that
gives access to see/decrypt certain fields

These two operations are extremely similar, but we had separate code for
checking both of them. Any permissions that were added needed to be
adjusted in multiple places.

With recent progress in `DataIdentifiers`, it becomes a lot more clear
to model these with the same code. Now, an `ObConiguration` maps to a
list of `TenantScope`s that specify the fields whose existence is
_visible_ (aka were requested to be collected) and a list of
`TenantScope`s that specify which fields are decryptable.

We then can use the same `CanDecrypt` guard to check if the guard is met
by the scopes on the ob config!

## Caveats

Now, the name `CanDecrypt` is a bit of a misnomer, because we sometimes
use `TenantScopes` that just specify whether a field's existence is
visible (not whether it is decrytpable). I still think this is cleaner
than before, but we can tweak names. Maybe to `CanSee`?